% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate1}
\alias{calibrate1}
\title{Row-calibration of log-parameter}
\usage{
calibrate1(theta, n, w_j.plus)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w_j.plus}{Row-sums of sample weights}
}
\value{
Row-calibrated log-parameter
}
\description{
Row-calibration of log-parameter
}
