% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate1_cpp}
\alias{calibrate1_cpp}
\title{Row-calibration of log-parameter}
\usage{
calibrate1_cpp(theta, n, w_jplus)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w_jplus}{Row-sums of sample weights}
}
\value{
Row-calibrated log-parameter
}
\description{
Row-calibration of log-parameter
}
