% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate2_cpp}
\alias{calibrate2_cpp}
\title{Column-calibration of log-parameter}
\usage{
calibrate2_cpp(theta, n, w_plusk)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w_plusk}{Column-sums of sample weights}
}
\value{
Row-calibrated log-parameter
}
\description{
Column-calibration of log-parameter
}
