% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate_cpp}
\alias{calibrate_cpp}
\title{Approximate calibration of log-parameter}
\usage{
calibrate_cpp(theta, l, mM, n, w, w_jplus, w_plusk, prec)
}
\arguments{
\item{theta}{Log-parameter}

\item{l}{Number of unique covariates}

\item{mM}{(m_j,M_j) index pairs}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{w_jplus}{Row-sums of sample weights}

\item{w_plusk}{Column sums of sample weights}

\item{prec}{Precision for the approximate calibration}
}
\value{
Approximately calibrated log-parameter
}
\description{
Computes an optimal additive adjustment of theta, i.e. a matrix
theta.new with values
theta.new_jk = theta_jk + x_j + y_k
such that f(theta.new) is minimal.
}
