% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{CRPS}
\alias{CRPS}
\title{Computes the 'CRPS'}
\usage{
CRPS(x0, l0, res, a, b, rel.tol = 1e-09)
}
\arguments{
\item{x0}{Extended set of covariates}

\item{l0}{Cardinality of extended set of covariates \code{x0}}

\item{res}{A list containing \code{CDF_LR}, \code{CDF_ST}, \code{CDF_EMP} and \code{par}. I.e.,
the estimated conditional cumulative distribution functions under
likelihood-ratio order and usual stochastic order, the conditional empirical
distribution functions, all of which having \code{x0} for covariates, and a list
of parameters, of which the unique observations \code{y} and their number \code{m} are
relevant}

\item{a, b}{Shape and scale functions of the true Gamma conditional
distribution functions}

\item{rel.tol}{Relative tolerance for numerical integration}
}
\value{
A l0-by-3 matrix of conditional CRPS
}
\description{
Computes the 'CRPS'
}
