% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_reg_C}
\alias{dist_reg_C}
\title{Isotonic distributional regression (LR, ST, EMP), C++ version}
\usage{
dist_reg_C(X, Y, W, delta0, x0, ST = FALSE)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified sample weights}

\item{delta0}{Threshhold}

\item{x0}{Set of covariates on which to estimate the distributions}

\item{ST}{Boolean indicating whether or not the classical isotonic
distributional regression will also be computed}
}
\value{
Isotonic distributional regression(s) and estimation parameters
}
\description{
Isotonic distributional regression (LR, ST, EMP), C++ version
}
