% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_theta_cpp}
\alias{f_theta_cpp}
\title{Negative log-likelihood in terms of log-parameter, C++ version}
\usage{
f_theta_cpp(theta, l, n, mM, w)
}
\arguments{
\item{theta}{Log-parameter}

\item{l}{Number of unique covariates}

\item{n}{Sample size}

\item{mM}{(m_j,M_j) index pairs}

\item{w}{Sample weights}
}
\value{
Negative log-likelihood in terms of log-parameter
}
\description{
Negative log-likelihood in terms of log-parameter, C++ version
}
