% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{f_theta_R}
\alias{f_theta_R}
\title{Negative log-likelihood in terms of log-parameter, R version}
\usage{
f_theta_R(theta, n, w, PP)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{PP}{Reduced index space}
}
\value{
Negative log-likelihood in terms of log-parameter
}
\description{
Negative log-likelihood in terms of log-parameter, R version
}
