% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{ftheta_R}
\alias{ftheta_R}
\title{Negative log-likelihood in terms of log-parameter, R version}
\usage{
ftheta_R(theta, n, w, PP)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{PP}{Reduced index space}
}
\value{
Negative log-likelihood in terms of log-parameter
}
\description{
Negative log-likelihood in terms of log-parameter, R version
}
