% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Linear interpolation of the cdf's}
\usage{
interpolate(x0, x, CDF)
}
\arguments{
\item{x0}{Set of covariates on which to extend CDF}

\item{x}{Set of covariates of CDF}

\item{CDF}{Step-function matrix, its j-th row contains the cdf for X = x_j}
}
\value{
Linear interpolation of the cdf's on the new set x0
}
\description{
Linear interpolation of the cdf's
}
