% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_cpp}
\alias{interpolate_cpp}
\title{Linear interpolation of the cdf's, C++ version}
\usage{
interpolate_cpp(x0, x, CDF)
}
\arguments{
\item{x0}{Set of covariates on which to extend CDF}

\item{x}{Set of covariates of CDF, {X_1, X_2, ..., X_N}}

\item{CDF}{Step-function matrix, its j-th row contains the cdf for X = x_j}
}
\value{
Linear interpolation of the cdf's on the new set x0
}
\description{
Linear interpolation of the cdf's, C++ version
}
