% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{lambda1.to.theta}
\alias{lambda1.to.theta}
\title{Transforms lambda (row) into theta, R version}
\usage{
lambda1.to.theta(lambda, ell, m, mM)
}
\arguments{
\item{lambda}{Row-wise differences}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{mM}{(m_j,M_j) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (row) into theta, R version
}
