% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda1_to_theta_C}
\alias{lambda1_to_theta_C}
\title{Transforms lambda (row) into theta, C++ version}
\usage{
lambda1_to_theta_C(lambda, l, m, mM)
}
\arguments{
\item{lambda}{Row-wise differences}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{mM}{(m_j,M_j) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (row) into theta, C++ version
}
