% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda1_to_theta_cpp}
\alias{lambda1_to_theta_cpp}
\title{Transforms lambda (row) into theta}
\usage{
lambda1_to_theta_cpp(lambda, ell, m, mM)
}
\arguments{
\item{lambda}{Row-wise differences}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{mM}{(m_j,M_j) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (row) into theta
}
