% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{lambda2.to.theta}
\alias{lambda2.to.theta}
\title{Transforms lambda (column) into theta, R version}
\usage{
lambda2.to.theta(lambda, ell, m, lL)
}
\arguments{
\item{lambda}{Column-wise differences}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{lL}{(l_k,L_k) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (column) into theta, R version
}
