% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda2_to_theta_C}
\alias{lambda2_to_theta_C}
\title{Transforms lambda (column) into theta, C++ version}
\usage{
lambda2_to_theta_C(lambda, l, m, lL)
}
\arguments{
\item{lambda}{Column-wise differences}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{lL}{(l_k,L_k) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (column) into theta, C++ version
}
