% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda2_to_theta_cpp}
\alias{lambda2_to_theta_cpp}
\title{Transforms lambda (column) into theta}
\usage{
lambda2_to_theta_cpp(lambda, ell, m, lL)
}
\arguments{
\item{lambda}{Column-wise differences}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{lL}{(l_k,L_k) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (column) into theta
}
