% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{lambda2_to_theta_R}
\alias{lambda2_to_theta_R}
\title{Transforms lambda (column) into theta, R version}
\usage{
lambda2_to_theta_R(lambda, l, m, lL)
}
\arguments{
\item{lambda}{Column-wise differences}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{lL}{(l_k,L_k) index pairs}
}
\value{
Transformed parameter
}
\description{
Transforms lambda (column) into theta, R version
}
