% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{local.search1}
\alias{local.search1}
\title{Local search (row)}
\usage{
local.search1(theta, ell, m, n, mM, lL, PP, w, w_cumul.1)
}
\arguments{
\item{theta}{Log-parameter}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{mM}{(m_j,M_j) index pairs}

\item{lL}{(l_k,L_k) index pairs}

\item{PP}{Reduced index space}

\item{w}{Sample weights}

\item{w_cumul.1}{Cumulative row-sums of sample weights}
}
\description{
Local search (row)
}
