% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{local_search1_C}
\alias{local_search1_C}
\title{Local search (row), C++ version}
\usage{
local_search1_C(theta, l, m, n, lL, mM, w, w_ul)
}
\arguments{
\item{theta}{Log-parameter}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{lL}{(l_k,L_k) index pairs}

\item{mM}{(m_j,M_j) index pairs}

\item{w}{Sample weights}

\item{w_ul}{Cumulative row-sums of sample weights}
}
\value{
New proposal Psi and step-size delta
}
\description{
Local search (row), C++ version
}
