% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotD}
\alias{plotD}
\title{Plot the design}
\usage{
plotD(par, indices = FALSE)
}
\arguments{
\item{par}{Model parameters}

\item{indices}{Whether or not to plot indices instead of true values}
}
\value{
Plot
}
\description{
Plot the design
}
