% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare.data}
\alias{prepare.data}
\title{Prepare the data}
\usage{
prepare.data(X, Y, W = rep(1, length(X)))
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified weights}
}
\value{
A list of pre-compute parameters necessary for the estimation
}
\description{
Prepare the data
}
