% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prepare_data_C}
\alias{prepare_data_C}
\title{Prepare the data, C++ version}
\usage{
prepare_data_C(X, Y, W)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified weights}
}
\value{
A list of pre-computed parameters necessary for the estimation
}
\description{
Prepare the data, C++ version
}
