% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prepare_data_cpp}
\alias{prepare_data_cpp}
\title{Prepare the data, C++ version}
\usage{
prepare_data_cpp(X, Y, W)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified weights}
}
\description{
Prepare the data, C++ version
}
