% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simple_step_cpp}
\alias{simple_step_cpp}
\title{Function to take a simple step}
\usage{
simple_step_cpp(theta, Psi, delta, l, mM, n, w)
}
\arguments{
\item{theta}{Log-parameter}

\item{Psi}{Proposal}

\item{delta}{Delta}

\item{l}{Number of unique covariates}

\item{mM}{(m_j,M_j) index pairs}

\item{n}{Sample size}

\item{w}{Sample weights}
}
\value{
Updated theta parameter
}
\description{
Function to take a simple step
}
