% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{simple_step_R}
\alias{simple_step_R}
\title{Function to take a simple step, R version}
\usage{
simple_step_R(theta, Psi, delta, l, m, n, w, PP)
}
\arguments{
\item{theta}{Log-parameter}

\item{Psi}{Proposal}

\item{delta}{Delta}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{PP}{Reduced index space}
}
\value{
Updated theta parameter
}
\description{
Function to take a simple step, R version
}
