% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taut_string.R
\name{taut.string}
\alias{taut.string}
\title{Taut string}
\usage{
taut.string(Y.low, Y.upp, x = 1:length(Y.low))
}
\arguments{
\item{Y.low}{Lower vector}

\item{Y.upp}{Upper vector}

\item{x}{Vector of support points}
}
\value{
Taut string between Y.low and Y.upp
}
\description{
The three vectors Y.low, Y.upp and x of length n > 2 are such
that
Y.low.i <= Y.upp.i  for all i
with equality for i in {1,n}, and
x1 < x2 < ... < xn.
The function outputs the corresponding taut string Y. That means, Y is a
vector of length n such that
Y.low.i <= Y.i <= Y.upp.i  for all i,
for 1 < i < n,
Y.i = Y.low.i  if  dY.i-1/dx.i-1 > dY.i/dx.i ,
Y.i = Y.upp.i  if  dY.i-1/dx.i-1 < dY.i/dx.i ,
where dz.j := z.j+1 - z.j for any vector z and 1 <= j < length(z).
}
