% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taut_string.R
\name{taut_string}
\alias{taut_string}
\title{Taut string}
\usage{
taut_string(Y_low, Y_upp, x = 1:length(Y_low))
}
\arguments{
\item{Y_low}{Lower vector}

\item{Y_upp}{Upper vector}

\item{x}{Vector of support points}
}
\value{
Taut string between Y_low and Y_upp
}
\description{
The three vectors Y_low, Y_upp and x of length n > 2 are such
that
Y_low_i <= Y_upp_i  for all i
with equality for i in {1,n}, and
x1 < x2 < ... < xn.
The function outputs the corresponding taut string Y. That means, Y is a
vector of length n such that
Y_low_i <= Y_i <= Y_upp_i  for all i,
for 1 < i < n,
Y_i = Y_low_i  if  dY_i-1/dx_i-1 > dY_i/dx_i ,
Y_i = Y_upp_i  if  dY_i-1/dx_i-1 < dY_i/dx_i ,
where dz_j := z_j+1 - z_j for any vector z and 1 <= j < length(z).
}
