% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TP2.fit}
\alias{TP2.fit}
\title{TP2 fit function}
\usage{
TP2.fit(D.Setup, delta0 = 0.1, echo = FALSE, out.file = FALSE)
}
\arguments{
\item{D.Setup}{Parameters}

\item{delta0}{Threshhold}

\item{echo}{Whether or not to print intermediate steps}

\item{out.file}{Whether or not to wwite intermediate steps in a separate file}
}
\description{
TP2 fit function
}
