% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TP2_fit_bag_cpp}
\alias{TP2_fit_bag_cpp}
\title{TP2 fit function, with bagging}
\usage{
TP2_fit_bag_cpp(X, Y, W, delta0, n_boot)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified weights}

\item{delta0}{Threshhold}

\item{n_boot}{Number of boostrap samples}
}
\value{
h matrix, delta and parameters (estimation time)
}
\description{
TP2 fit function, with bagging
}
