% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TP2_fit_cpp}
\alias{TP2_fit_cpp}
\title{TP2 fit function}
\usage{
TP2_fit_cpp(X, Y, W, delta0)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified weights}

\item{delta0}{Threshhold}
}
\value{
h matrix, delta and parameters (estimation time)
}
\description{
TP2 fit function
}
