% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TP2_fit_R}
\alias{TP2_fit_R}
\title{TP2 fit function, R version}
\usage{
TP2_fit_R(par, delta0 = 1e-08)
}
\arguments{
\item{par}{Model parameters}

\item{delta0}{Threshold}
}
\value{
h matrix and estimation time
}
\description{
TP2 fit function, R version
}
