% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vg_tilde2_C}
\alias{vg_tilde2_C}
\title{v-tilde and gamma-tilde functions (column), C++ version}
\usage{
vg_tilde2_C(theta, l, m, n, lL, w_ol)
}
\arguments{
\item{theta}{Log-parameter}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{lL}{(l_k,L_k) index pairs}

\item{w_ol}{Cumulative column-sums of sample weights}
}
\value{
v-tilde and gamma-tilde functions
}
\description{
v-tilde and gamma-tilde functions (column), C++ version
}
