% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{vgamma.tilde1}
\alias{vgamma.tilde1}
\title{v-tilde and gamma-tilde functions (row)}
\usage{
vgamma.tilde1(theta, ell, m, n, mM, w_cumul.1)
}
\arguments{
\item{theta}{Log-parameter}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{mM}{(m_j,M_j) index pairs}

\item{w_cumul.1}{Cumulative row-sums of sample weights}
}
\description{
v-tilde and gamma-tilde functions (row)
}
