% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{vgamma.tilde2}
\alias{vgamma.tilde2}
\title{v-tilde and gamma-tilde functions (column)}
\usage{
vgamma.tilde2(theta, ell, m, n, lL, w_ol)
}
\arguments{
\item{theta}{Log-parameter}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{lL}{(l_k,L_k) index pairs}

\item{w_ol}{Cumulative column-sums of sample weights}
}
\value{
v-tilde and gamma-tilde functions
}
\description{
v-tilde and gamma-tilde functions (column)
}
