% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vgamma_tilde1_cpp}
\alias{vgamma_tilde1_cpp}
\title{v-tilde and gamma-tilde functions (row), C++ version}
\usage{
vgamma_tilde1_cpp(theta, l, m, n, mM, w_ul)
}
\arguments{
\item{theta}{Log-parameter}

\item{l}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{mM}{(m_j,M_j) index pairs}

\item{w_ul}{Cumulative row-sums of sample weights}
}
\value{
v-tilde and gamma-tilde functions
}
\description{
v-tilde and gamma-tilde functions (row), C++ version
}
