% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Approximate calibration of log-parameter}
\usage{
calibrate(theta, n, w, w_jplus, w_plusk, PP, prec = 1e-10)
}
\arguments{
\item{theta}{Log-parameter}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{w_jplus}{Row-sums of sample weights}

\item{w_plusk}{Column sums of sample weights}

\item{PP}{Reduced index space}

\item{prec}{Precision for the approximate calibration}
}
\value{
Approximately calibrated log-parameter
}
\description{
Computes an optimal additive adjustment of theta, i.e. a matrix
theta.new with values
theta.new_jk = theta_jk + x_j + y_k
such that f(theta.new) is minimal.
}
