% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dist.reg}
\alias{dist.reg}
\title{Isotonic distributional regression (LR, US, EMP)}
\usage{
dist.reg(
  X,
  Y,
  W = rep(1, length(X)),
  show.design = FALSE,
  indices = FALSE,
  suggest.delta0 = FALSE,
  delta0 = 0.01,
  echo = FALSE,
  out.file = FALSE,
  IDR = FALSE,
  x0 = NULL
)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Responses}

\item{W}{User-specified sample weights}

\item{show.design}{Whether or not to plot the design}

\item{indices}{Whether or not to plot indices instead of true values}

\item{suggest.delta0}{Whether or not to suggest the threshold delta0}

\item{delta0}{Threshhold}

\item{echo}{Whether or not to print intermediate steps}

\item{out.file}{Whether or not to wwite intermediate steps in a separate file}

\item{IDR}{Whether or not to compute Isotonic Distributional Regression}

\item{x0}{Set of covariates on which to estimate the distributions}
}
\value{
A list of results which depends on the option chosen
}
\description{
Isotonic distributional regression (LR, US, EMP)
}
\examples{
#To be done
}
