% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Title}
\usage{
interpolate(x0, x, CDF.MAT)
}
\arguments{
\item{x0}{Set of covariates on which to extend CDF.MAT}

\item{x}{Set of covariates of CDF.MAT, {X_1, X_2, ..., X_N}}

\item{CDF.MAT}{Step-function matrix, its j-th row contains the cdf for X = x_j}
}
\description{
Title
}
