% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{local.search2}
\alias{local.search2}
\title{Local search (column)}
\usage{
local.search2(theta, ell, m, n, mM, lL, PP, w, w_ol)
}
\arguments{
\item{theta}{Log-parameter}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{mM}{(m_j,M_j) index pairs}

\item{lL}{(l_k,L_k) index pairs}

\item{PP}{Reduced index space}

\item{w}{Sample weights}

\item{w_ol}{Cumulative column-sums of sample weights}
}
\value{
New proposal Psi and step-size delta
}
\description{
Local search (column)
}
