% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{simple.step}
\alias{simple.step}
\title{Function to take a simple step}
\usage{
simple.step(theta, Psi, delta, ell, m, n, w, PP)
}
\arguments{
\item{theta}{Log-parameter}

\item{Psi}{Proposal}

\item{delta}{Delta}

\item{ell}{Number of unique covariates}

\item{m}{Number of unique responses}

\item{n}{Sample size}

\item{w}{Sample weights}

\item{PP}{Reduced index space}
}
\value{
Updated theta parameter
}
\description{
Function to take a simple step
}
