% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TP2.fit}
\alias{TP2.fit}
\title{TP2 fit function}
\usage{
TP2.fit(par, delta0 = 0.1)
}
\arguments{
\item{par}{Model parameters}

\item{delta0}{Threshold}
}
\value{
h matrix and estimation time
}
\description{
TP2 fit function
}
